nautilus_coinbase_intx::common

Module enums

Structs

CoinbaseIntxAlgoStrategyIter

An iterator over the variants of CoinbaseIntxAlgoStrategy

CoinbaseIntxAssetStatusIter

An iterator over the variants of CoinbaseIntxAssetStatus

CoinbaseIntxBookActionIter

An iterator over the variants of CoinbaseIntxBookAction

CoinbaseIntxCandleConfirmIter

An iterator over the variants of CoinbaseIntxCandleConfirm

CoinbaseIntxExecTypeIter

An iterator over the variants of CoinbaseIntxExecType

CoinbaseIntxFeeTierTypeIter

An iterator over the variants of CoinbaseIntxFeeTierType

CoinbaseIntxInstrumentTypeIter

An iterator over the variants of CoinbaseIntxInstrumentType

CoinbaseIntxOrderEventTypeIter

An iterator over the variants of CoinbaseIntxOrderEventType

CoinbaseIntxOrderStatusIter

An iterator over the variants of CoinbaseIntxOrderStatus

CoinbaseIntxOrderTypeIter

An iterator over the variants of CoinbaseIntxOrderType

CoinbaseIntxSTPModeIter

An iterator over the variants of CoinbaseIntxSTPMode

CoinbaseIntxSideIter

An iterator over the variants of CoinbaseIntxSide

CoinbaseIntxTimeInForceIter

An iterator over the variants of CoinbaseIntxTimeInForce

CoinbaseIntxTradingStateIter

An iterator over the variants of CoinbaseIntxTradingState

Enums

CoinbaseIntxAlgoStrategy

Represents an algo strategy.

CoinbaseIntxAssetStatus

Represents an asset status on Coinbase International.

CoinbaseIntxBookAction

Represents the type of book action.

CoinbaseIntxCandleConfirm

Represents the possible states of an order throughout its lifecycle.

CoinbaseIntxExecType

Represents the type of execution that generated a trade.

CoinbaseIntxFeeTierType

Represents the type of book action.

CoinbaseIntxInstrumentType

Represents instrument types on Coinbase International.

CoinbaseIntxOrderEventType

Represents an order event type.

CoinbaseIntxOrderStatus

Represents an overall order status.

CoinbaseIntxOrderType

Represents an order type.

CoinbaseIntxSTPMode

Represents a self trade protection (STP) mode.

CoinbaseIntxSide

Represents the side of an order or trade (Buy/Sell).

CoinbaseIntxTimeInForce

Represents an order time in force.

CoinbaseIntxTradingState

Represents an instrument status on Coinbase International.