nautilus_coinbase_intx::common

Module parse

Functions

bar_spec_as_coinbase_channel

Converts a BarSpecification into the corresponding Coinbase WebSocket channel.

coinbase_channel_as_bar_spec

Converts a Coinbase WebSocket channel into the corresponding BarSpecification.

deserialize_optional_string_to_u64

Custom deserializer for strings to u64.

get_currency

Returns the currency either from the internal currency map or creates a default crypto. Returns the currency either from the internal currency map or creates a default crypto.

parse_aggressor_side

parse_execution_type

parse_instrument_id

Parses a Posei instrument ID from the given Coinbase symbol value.

parse_millisecond_timestamp

Parses a timestamp in milliseconds since epoch into UnixNanos.

parse_notional

Parses a notional string into Money, returning None if the value is zero.

parse_order_side

parse_position_side

parse_price

Parses a string into a Price.

parse_quantity

Parses a string into a Quantity with the given precision.

parse_rfc3339_timestamp

Parses an RFC3339 timestamp string into UnixNanos.