nautilus_coinbase_intx::http

Module client

Provides the HTTP client integration for the Coinbase International REST API.

This module defines and implements a CoinbaseIntxHttpClient for sending requests to various Coinbase endpoints. It handles request signing (when credentials are provided), constructs valid HTTP requests using the [HttpClient], and parses the responses back into structured data or a CoinbaseIntxHttpError.

Structs

CoinbaseIntxHttpClient

Provides a higher-level HTTP client for the Coinbase International REST API.

CoinbaseIntxHttpInnerClient

Provides a lower-level HTTP client for connecting to the Coinbase International REST API.

CoinbaseIntxResponse

Represents an Coinbase HTTP response.

Statics

COINBASE_INTX_REST_QUOTA