nautilus_coinbase_intx::http::client

Struct CoinbaseIntxHttpClient

pub struct CoinbaseIntxHttpClient { /* private fields */ }

Provides a higher-level HTTP client for the Coinbase International REST API.

This client wraps the underlying CoinbaseIntxHttpInnerClient to handle conversions into the Posei domain model.

Implementations

impl CoinbaseIntxHttpClient

pub fn new(base_url: Option<String>, timeout_secs: Option<u64>) -> Self

Creates a new CoinbaseIntxHttpClient using the default Coinbase HTTP URL, optionally overridden with a custom base url.

This version of the client has no credentials, so it can only call publicly accessible endpoints.

pub fn from_env() -> Result<Self>

Creates a new authenticated CoinbaseIntxHttpClient using environment variables and the default Coinbase International HTTP base url.

Errors

Returns an error if required environment variables are missing or invalid.

pub fn with_credentials( api_key: Option<String>, api_secret: Option<String>, api_passphrase: Option<String>, base_url: Option<String>, timeout_secs: Option<u64>, ) -> Result<Self>

Creates a new CoinbaseIntxHttpClient configured with credentials for authenticated requests, optionally using a custom base url.

Errors

Returns an error if required environment variables are missing or invalid.

pub fn base_url(&self) -> &str

Returns the base url being used by the client.

pub fn api_key(&self) -> Option<&str>

Returns the public API key being used by the client.

pub const fn is_initialized(&self) -> bool

Checks if the client is initialized.

The client is considered initialized if any instruments have been cached from the venue.

pub fn get_cached_symbols(&self) -> Vec<String>

Returns the cached instrument symbols.

Panics

Panics if the instrument cache mutex is poisoned.

pub fn add_instruments(&mut self, instruments: Vec<InstrumentAny>)

Adds the given instruments into the clients instrument cache.

Panics

Panics if the instrument cache mutex is poisoned.

Any existing instruments will be replaced.

pub fn add_instrument(&mut self, instrument: InstrumentAny)

Adds the given instrument into the clients instrument cache.

Panics

Panics if the instrument cache mutex is poisoned.

Any existing instrument will be replaced.

pub async fn list_portfolios(&self) -> Result<Vec<CoinbaseIntxPortfolio>>

Requests a list of portfolio details from Coinbase International.

Errors

Returns an error if the HTTP request fails or the response cannot be parsed.

pub async fn request_account_state( &self, account_id: AccountId, ) -> Result<AccountState>

Requests the account state for the given account ID from Coinbase International.

Errors

Returns an error if the HTTP request fails or the response cannot be parsed.

pub async fn request_instruments(&self) -> Result<Vec<InstrumentAny>>

Requests all instruments from Coinbase International.

Errors

Returns an error if the HTTP request fails or the response cannot be parsed.

pub async fn request_instrument(&self, symbol: &Symbol) -> Result<InstrumentAny>

Requests the instrument for the given symbol from Coinbase International.

Errors

Returns an error if the HTTP request fails or the instrument cannot be parsed.

pub async fn request_order_status_report( &self, account_id: AccountId, venue_order_id: VenueOrderId, ) -> Result<OrderStatusReport>

Requests an order status report for the given venue order ID from Coinbase International.

Errors

Returns an error if the HTTP request fails or the response cannot be parsed.

pub async fn request_order_status_reports( &self, account_id: AccountId, symbol: Symbol, ) -> Result<Vec<OrderStatusReport>>

Requests order status reports for all open orders from Coinbase International.

Errors

Returns an error if the HTTP request fails or the response cannot be parsed.

pub async fn request_fill_reports( &self, account_id: AccountId, client_order_id: Option<ClientOrderId>, start: Option<DateTime<Utc>>, ) -> Result<Vec<FillReport>>

Requests all fill reports from Coinbase International.

Errors

Returns an error if the HTTP request fails or the response cannot be parsed.

pub async fn request_position_status_report( &self, account_id: AccountId, symbol: Symbol, ) -> Result<PositionStatusReport>

Requests a position status report from Coinbase International.

Errors

Returns an error if the HTTP request fails or the response cannot be parsed.

pub async fn request_position_status_reports( &self, account_id: AccountId, ) -> Result<Vec<PositionStatusReport>>

Requests all position status reports from Coinbase International.

Errors

Returns an error if the HTTP request fails or the response cannot be parsed.

pub async fn submit_order( &self, account_id: AccountId, client_order_id: ClientOrderId, symbol: Symbol, order_side: OrderSide, order_type: OrderType, quantity: Quantity, time_in_force: TimeInForce, expire_time: Option<DateTime<Utc>>, price: Option<Price>, trigger_price: Option<Price>, post_only: Option<bool>, reduce_only: Option<bool>, ) -> Result<OrderStatusReport>

Submits a new order to Coinbase International.

Errors

Returns an error if the HTTP request fails or the response cannot be parsed.

pub async fn cancel_order( &self, account_id: AccountId, client_order_id: ClientOrderId, ) -> Result<OrderStatusReport>

Cancels a currently open order on Coinbase International.

Errors

Returns an error if the HTTP request fails or the response cannot be parsed.

pub async fn cancel_orders( &self, account_id: AccountId, symbol: Symbol, order_side: Option<OrderSide>, ) -> Result<Vec<OrderStatusReport>>

Cancels all orders for the given account ID and filter params on Coinbase International.

Errors

Returns an error if the HTTP request fails or the response cannot be parsed.

pub async fn modify_order( &self, account_id: AccountId, client_order_id: ClientOrderId, new_client_order_id: ClientOrderId, price: Option<Price>, trigger_price: Option<Price>, quantity: Option<Quantity>, ) -> Result<OrderStatusReport>

Modifies a currently open order on Coinbase International.

Errors

Returns an error if the HTTP request fails or the response cannot be parsed.

impl CoinbaseIntxHttpClient

pub fn py_base_url(&self) -> &str

pub fn py_api_key(&self) -> Option<&str>

pub const fn py_is_initialized(&self) -> bool

pub fn py_get_cached_symbols(&self) -> Vec<String>

pub fn py_add_instrument( &mut self, py: Python<'_>, instrument: PyObject, ) -> PyResult<()>

Errors

Returns a Python exception if adding the instrument to the cache fails.

pub fn py_list_portfolios<'py>( &mut self, py: Python<'py>, ) -> PyResult<Bound<'py, PyAny>>

Errors

Returns a Python exception if retrieving or serializing portfolios fails.