nautilus_coinbase_intx::http

Module parse

Functions

parse_account_state

Parses account balances into an AccountState.

parse_fill_report

Parses a fill report from raw Coinbase REST data.

parse_instrument_any

parse_order_status_report

Parses an order status report from raw Coinbase REST data.

parse_perp_instrument

Parses a Coinbase International perpetual instrument into an InstrumentAny::CryptoPerpetual. Parses a perpetual instrument definition into an InstrumentAny::CryptoPerpetual.

parse_position_status_report

Parses a position status report from raw Coinbase REST data.

parse_spot_instrument

Parses a Coinbase International Spot instrument into an InstrumentAny::CurrencyPair. Parses a spot instrument definition into an InstrumentAny::CurrencyPair.