nautilus_coinbase_intx::http::parse

Function parse_spot_instrument

pub fn parse_spot_instrument(
    definition: &CoinbaseIntxInstrument,
    margin_init: Option<Decimal>,
    margin_maint: Option<Decimal>,
    maker_fee: Option<Decimal>,
    taker_fee: Option<Decimal>,
    ts_init: UnixNanos,
) -> Result<InstrumentAny>

Parses a Coinbase International Spot instrument into an InstrumentAny::CurrencyPair. Parses a spot instrument definition into an InstrumentAny::CurrencyPair.

Errors

Returns an error if any numeric field cannot be parsed or required data is missing.